backend/app/auth/jwt.py
from datetime import UTC, datetime, timedelta
from typing import Any

from jose import JWTError, jwt

from app.utils.config import get_settings

settings = get_settings()


def create_access_token(subject: str, additional_claims: dict[str, Any] | None = None) -> str:
    expire_at = datetime.now(UTC) + timedelta(minutes=settings.ACCESS_TOKEN_EXPIRE_MINUTES)
    payload: dict[str, Any] = {"sub": subject, "exp": expire_at}
    if additional_claims:
        payload.update(additional_claims)
    return jwt.encode(payload, settings.JWT_SECRET_KEY, algorithm=settings.JWT_ALGORITHM)


def decode_access_token(token: str) -> dict[str, Any] | None:
    try:
        return jwt.decode(token, settings.JWT_SECRET_KEY, algorithms=[settings.JWT_ALGORITHM])
    except JWTError:
        return None
