backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pydantic==2.10.4
pydantic-settings==2.7.0
email-validator==2.2.0
python-multipart==0.0.17
alembic==1.14.0
